Use incremental snapshots from clones instead of zfs promote trickery
Image format version is increased from z0 to z1.

Switched image transfer from an incremental snapshot on a single
filesystem to an incremental snapshot from a cloned filesystem. This is
a backward incompatible change in saved images and in images in
repository.

The format of the images stored locally was not changed, but scratch
image is handled differently, and it is cleaner to clean all images from
the host and the old scratch image manually.

The change simplifies ZFS commands and removes inconsistencies that were
related to terminated or simultaneous executions. There is no longer need
to implement locking for these critical parts either.